updated login page and edited styling
