+++ config/connection.js
@@ -8,7 +8,8 @@ const sequelize = process.env.JAWSDB_URL
   : new Sequelize(process.env.DB_NAME, process.env.DB_USER, process.env.DB_PW, {
       host: 'localhost',
       dialect: 'mysql',
-      port: 3306
+      port: 3306,
+      logging: false
     });
 
 module.exports = sequelize;
