+++ controllers/home-routes.js
@@ -1,13 +1,32 @@
 const { Game, Categories } = require('../models');
-
+const sequelize = require("../config/connection");
 const router = require('express').Router();
 
 // When the user navigates to the home page they are sent to the index.js file which routes them to this file
 // Then this file renders the homepage.handlebars template
 router.get('/', (req, res) => {
+
+  if(req.session.loggedIn) {
   res.render('homepage',{
     title: 'Home Page',
-    style: 'main.css',
+    loggedIn: req.session.loggedIn
+  });
+} else {
+  res.redirect('/');
+  return;
+}
+});
+
+router.get('/signup', (req, res) => {
+
+  if (req.session.loggedIn) {
+    res.redirect('/');
+    return;
+  }
+
+  res.render('signup',{
+    title: 'Sign Up',
+    loginPage: false,
   });
 });
 
@@ -19,7 +38,7 @@ router.get('/login', (req, res) => {
 
   res.render('login',{
     title: 'Login',
-    style: 'login.css'
+    loginPage: true,
   });
 });
