+++ public/javascript/login.js
@@ -22,36 +22,8 @@ async function loginFormHandler(event) {
     }
   }
   
-  async function signupFormHandler(event) {
-    event.preventDefault();
-  
-    const username = document.querySelector("#username-signup").value.trim();
-    const email = document.querySelector("#email-signup").value.trim();
-    const password = document.querySelector("#password-signup").value.trim();
-  
-    if (username && email && password) {
-      const response = await fetch("/api/users", {
-        method: "post",
-        body: JSON.stringify({
-          username,
-          email,
-          password,
-        }),
-        headers: { "Content-Type": "application/json" },
-      });
-  
-      if (response.ok) {
-        document.location.replace("/profile");
-      } else {
-        alert(response.statusText);
-      }
-    }
-  }
-  
+
   document
     .querySelector(".login-form")
     .addEventListener("submit", loginFormHandler);
   
-  document
-    .querySelector(".signup-form")
-    .addEventListener("submit", signupFormHandler);
