+++ public/stylesheet/login.css
@@ -1,100 +1,8 @@
-* {
-	box-sizing: border-box;
+.btn-danger {
+  background-color: #ff0000;
+  font-size: 2.5vw;
 }
 
-.header {
-	padding: 30px;
-	text-align: center;
-	background:#C70000;
-	color: white;
-  }
-
-nav {
-	overflow: hidden;
-	background-color: #8FBC72;
-	border: 1px solid black;
-  }
-  
-nav a {
-	float: left;
-	display: block;
-	color:black;
-	text-align: center;
-	padding: 14px 16px;
-	text-decoration: none;
-	border-right: 1px solid black;
-  }
-   
-.column {
-    float: left;
-    width: 50%;
-    }
-
-.row:after {
-    content: "";
-    display: table;
-    clear: both;
-    }
-
-body {
-	background: #BFDBAE;
+.form-group {
+	font-size: 2vw;;
 }
-
-h1 {
-	font-weight: bold;
-	margin: 0;
-}
-
-#sign-in, #sign-up {
-	border-radius: 20px;
-	border: 1px solid #FF4B2B;
-	background-color: #FF4B2B;
-	color: #FFFFFF;
-	font-size: 12px;
-	font-weight: bold;
-	padding: 12px 45px;
-	letter-spacing: 1px;
-}
-
-form {
-	background-color: #FFFFFF;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	flex-direction: column;
-	padding: 0 50px;
-	height: 100%;
-	text-align: center;
-}
-
-input {
-	background-color: #eee;
-	border: none;
-	padding: 12px 15px;
-	margin: 8px 0;
-	width: 100%;
-}
-
-.josh {
-	background-color: orange;
-	width: 90px;
-	height: 40px;
-  }
-  
-  .mark {
-	background-color: green;
-	width: 90px;
-	height: 40px;
-  }
-  
-  .jessica {
-	background-color: yellow;
-	width: 90px;
-	height: 40px;
-  }
-  
-  .jane {
-	background-color: magenta;
-	width: 90px;
-	height: 40px;
-  }
