+++ server.js
@@ -3,6 +3,7 @@ const express = require('express');
 const fileUpload = require('express-fileupload')
 const session = require('express-session');
 const exphbs = require('express-handlebars');
+const routes = require('./controllers');
 
 const app = express();
 const PORT = process.env.PORT || 3001;
@@ -33,7 +34,7 @@ app.use(express.urlencoded({ extended: false }));
 app.use(express.static(path.join(__dirname, 'public')));
 
 app.use(fileUpload());
-app.use(require('./controllers/'));
+app.use(routes);
 
 sequelize.sync({ force: false }).then(() => {
   app.listen(PORT, () => console.log(`Now listening on ${PORT}`));
