+++ views/layouts/main.handlebars
@@ -1,81 +1,114 @@
-<!DOCTYPE html>
 <html lang="en">
 
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        {{!-- <link rel="stylesheet" href="/stylesheet/{{style}}" /> --}}
-        {{!-- bootstrap --}}
-        <link rel="stylesheet"href="https://cdnjs.cloudflare.com/ajax/libs/bootswatch/4.1.3/flatly/bootstrap.min.css" />
-        <link rel="stylesheet" href="/stylesheet/main.css" />
-        <title>{{title}}</title>
-    </head>
-    <body>
-        {{!-- hero  and naviagion links--}}
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 
-        <header>
-            
-            <div class = " hero row">
-                <img src="/images/logo.png">
-            </div>
+    <link
+      rel="stylesheet"
+      href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"
+      integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm"
+      crossorigin="anonymous"
+    />
 
+    <script
+      src="https://code.jquery.com/jquery-3.2.1.slim.min.js"
+      integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN"
+      crossorigin="anonymous"
+    ></script>
+    <script
+      src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js"
+      integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q"
+      crossorigin="anonymous"
+    ></script>
+    <script
+      src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"
+      integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl"
+      crossorigin="anonymous"
+    ></script>
 
-            <div class = "nav ">
-                <div class = "col-4">
-                    <a href="/">Home</a>
-                </div>
-                <div class = "col-4">
-                    <a href="/game">Game</a>
-                </div>
-                <div class = "col-4">
-                    {{#if loggedIn}}
-                    <a href="/profile">Profile</a>
-                    <button id="logout" class="btn-no-style">Logout</button>
-                </div>
-                <div class = "col-4">
-                    {{else}}
-                    <a href="/login">Login</a>
-                    {{/if}}
-                </div>
-            </div>
-            
-        </header>
+    <link rel="stylesheet" href="/stylesheet/main.css" defer />
+    <link rel="stylesheet" href="/stylesheet/login.css" defer />
+    <title>{{title}}</title>
+  </head>
+  <body>
+    {{! hero  and naviagion links}}
 
-        {{!-- page content --}}
-        <main>
-            {{{body}}}
-        </main>
-        <footer class = "container">
-            <div class = "row justify-content-center">
-                <div class = "col-12">
-                    <h2><strong>Made by Group 3:</strong></h2>
-                </div>
-            </div>
-            <div class = "row justify-content-center">
-                
-                    <div class="github row">
-                        <!-- links to team's GitHub pages -->
-                        <div class="column">
-                            <a href="https://github.com/Joshsands" target="_blank"><button
-                                    class="josh">Joshua</button></a>
-                        </div>
-                        <div class="column">
-                            <a href="https://github.com/mjgiannelli" target="_blank"><button
-                                    class="mark">Mark</button></a>
-                        </div>
-                        <div class="column">
-                            <a href="https://github.com/JSheleg" target="_blank"><button
-                                    class="jessica">Jessica</button></a>
-                        </div>
-                        <div class="column">
-                            <a href="https://github.com/janekv20" target="_blank"><button
-                                    class="jane">Jane</button></a>
-                        </div>
-                    </div>
-                
-            </div>
-            
-            {{!-- <div class="content has-text-centered">
+    <header>
+      <div class="hero row">
+        <img src="/images/logo.png" />
+      </div>
+
+      <nav class="nav nav-pills justify-content-center">
+        <li class="nav-item">
+          <a class="nav-link btn btn-success" href="/">Home</a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link btn btn-success" href="/game">Game</a>
+        </li>
+        {{#if loggedIn}}
+          <li class="nav-item">
+            <a class="nav-link btn btn-success" href="/profile">Profile</a>
+          </li>
+          <li class="nav-item">
+            <a id="logout" class="nav-link btn btn-success">Logout</a>
+          </li>
+        {{else}}
+
+          {{#if loginPage}}
+            <li class="nav-item">
+              <a href="/signup" class="nav-link btn btn-success">Signup</a>
+            </li>
+          {{else}}
+            <li class="nav-item">
+              <a href="/login" class="nav-link btn btn-success">Login</a>
+            </li>
+          {{/if}}
+
+        {{/if}}
+      </nav>
+
+    </header>
+
+    {{! page content }}
+    <main>
+      {{{body}}}
+    </main>
+    <footer class="container">
+      <div class="row justify-content-center">
+        <div class="col-12">
+          <h2 class="text-white"><strong>Made by Group 3:</strong></h2>
+        </div>
+      </div>
+      <div class="row justify-content-center">
+
+        <div class="github row">
+          <!-- links to team's GitHub pages -->
+          <div class="column">
+            <a href="https://github.com/Joshsands" target="_blank"><button
+                class="josh"
+              >Joshua</button></a>
+          </div>
+          <div class="column">
+            <a href="https://github.com/mjgiannelli" target="_blank"><button
+                class="mark"
+              >Mark</button></a>
+          </div>
+          <div class="column">
+            <a href="https://github.com/JSheleg" target="_blank"><button
+                class="jessica"
+              >Jessica</button></a>
+          </div>
+          <div class="column">
+            <a href="https://github.com/janekv20" target="_blank"><button
+                class="jane"
+              >Jane</button></a>
+          </div>
+        </div>
+
+      </div>
+
+      {{! <div class="content has-text-centered">
                 <div class="github row">
                     <!-- links to team's GitHub pages -->
                     <div class="column">
@@ -95,16 +128,14 @@
                                 class="jane">Jane</button></a>
                     </div>
                 </div>
-            </div> --}}
+            </div> }}
 
-            
-        </footer>
-        
-        
-        {{#if loggedIn}}
-        <script src="/javascript/logout.js"></script>
-        {{/if}}     
-            
-    </body>
+    </footer>
+
+    {{#if loggedIn}}
+      <script src="/javascript/logout.js"></script>
+    {{/if}}
+
+  </body>
 
 </html>
