+++ views/signup.handlebars
@@ -1,0 +1,42 @@
+<br />
+<section class="w-25 mx-auto text-white">
+  <h1 class="text-center text-danger">Sign Up</h1>
+
+  <form class="signup-form">
+    <div class="form-group">
+      <label for="username-signup">Username:</label>
+      <input
+        type="text"
+        id="username-signup"
+        class="form-control"
+        placeholder="Username"
+      />
+    </div>
+
+    <div class="form-group">
+      <label for="email-signup">E-mail:</label>
+      <input
+        type="text"
+        id="email-signup"
+        class="form-control"
+        placeholder="Email"
+      />
+    </div>
+
+    <div class="form-group">
+      <label for="password-signup">Password:</label>
+      <input
+        type="password"
+        id="password-signup"
+        class="form-control"
+        placeholder="Password"
+      />
+    </div>
+    <div class="signup-form">
+      <button type="submit" class="btn btn-danger btn-block">Signup</button>
+    </div>
+  </form>
+</section>
+<br />
+
+<script src="/javascript/signup.js"></script>
